Created an action to set argument file
regrouped forms and views into forms.py and views.py
removed code from __init__ and added them as actions
