+++ actions/radicale
@@ -38,6 +38,8 @@ def parse_arguments():
     subparsers.add_parser('setup', help='Setup Radicale configuration')
     subparsers.add_parser('enable', help='Enable Radicale service')
     subparsers.add_parser('disable', help='Disable Radicale service')
+    configure = subparsers.add_parser('configure', help='configure various options')
+    configure.add_argument('--rights_type', help='Set the rights type for radicale')
 
     return parser.parse_args()
 
@@ -58,6 +60,16 @@ def subcommand_setup(_):
     aug.set('/files' + CONFIG_FILE + '/auth/type', 'remote_user')
     aug.set('/files' + CONFIG_FILE + '/rights/type', 'owner_only')
 
+    aug.save()
+
+    action_utils.service_restart('radicale')
+    action_utils.webserver_enable('radicale-plinth')
+
+
+def subcommand_configure(arguments):
+    """Sets the radicale rights type to a particular value"""
+    aug = load_augeas()
+    aug.set('/files' + CONFIG_FILE + '/rights/type', arguments.rights_type)
     aug.save()
 
     action_utils.service_restart('radicale')
